fix: ensure correct architecture build artifacts are downloaded
To ensure this I have also added much more verbose logging for instances in which a build's
architecture does not match the architecture queried - for example is `Architecture: all`

Another change to support this was when querying by source package we need to iterate over
all the published source packages for all releases until we find one with builds.

This is because versions are often copied from release to release without a rebuild.

This fixes issue https://github.com/canonical/ubuntu-package-buildinfo/issues/5