fix: Handle source package builds that are not for all architectures
This is because there exists a scenario where builds exist for the source
package in the specified series but not for the requested architecture. An example of where
this might occur is for introduction of a new architecture like riscv64 where only that
architecture is built for the source package as the other builds for the other architectures
have already been published.